requirements.txt
discord.py == 2.4.0
beautifulsoup4 == 4.12.3
wavelink == 3.4.0
pymongo == 4.8.0